Em-2945: Finish swagger docs and refactor to move them to documentation folder
